Refactor XCB code
